refactor: move common date input logic into base class (#18213)
Moves the common logic for a date input into a base class and implements it for the range inputs.
