+++ src/dev-app/datepicker/datepicker-demo.scss
@@ -2,3 +2,6 @@ mat-calendar {
   width: 300px;
 }
 
+.demo-range-group {
+  margin-bottom: 30px;
+}
