+++ src/material/datepicker/datepicker-input-base.ts
@@ -1,0 +1,286 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import {BooleanInput, coerceBooleanProperty} from '@angular/cdk/coercion';
+import {DOWN_ARROW} from '@angular/cdk/keycodes';
+import {
+  Directive,
+  ElementRef,
+  EventEmitter,
+  Inject,
+  Input,
+  OnDestroy,
+  Optional,
+  Output,
+} from '@angular/core';
+import {
+  AbstractControl,
+  ControlValueAccessor,
+  ValidationErrors,
+  Validator,
+  ValidatorFn,
+} from '@angular/forms';
+import {
+  DateAdapter,
+  MAT_DATE_FORMATS,
+  MatDateFormats,
+  MatDateSelectionModel,
+} from '@angular/material/core';
+import {Subscription} from 'rxjs';
+import {createMissingDateImplError} from './datepicker-errors';
+
+/**
+ * An event used for datepicker input and change events. We don't always have access to a native
+ * input or change event because the event may have been triggered by the user clicking on the
+ * calendar popup. For consistency, we always use MatDatepickerInputEvent instead.
+ */
+export class MatDatepickerInputEvent<D> {
+  /** The new value for the target datepicker input. */
+  value: D | null;
+
+  constructor(
+      /** Reference to the datepicker input component that emitted the event. */
+      public target: MatDatepickerInputBase<D>,
+      /** Reference to the native input element associated with the datepicker input. */
+      public targetElement: HTMLElement) {
+    this.value = this.target.value;
+  }
+}
+
+/** Base class for datepicker inputs. */
+@Directive()
+export abstract class MatDatepickerInputBase<D> implements ControlValueAccessor, OnDestroy,
+  Validator {
+  /** The value of the input. */
+  @Input()
+  get value(): D | null { return this._model ? this._model.selection : this._pendingValue; }
+  set value(value: D | null) {
+    value = this._dateAdapter.deserialize(value);
+    this._lastValueValid = !value || this._dateAdapter.isValid(value);
+    value = this._getValidDateOrNull(value);
+    const oldDate = this.value;
+    this._assignValue(value);
+    this._formatValue(value);
+
+    if (!this._dateAdapter.sameDate(oldDate, value)) {
+      this._valueChange.emit(value);
+    }
+  }
+  protected _model: MatDateSelectionModel<D | null, D> | undefined;
+
+  /** Whether the datepicker-input is disabled. */
+  @Input()
+  get disabled(): boolean { return !!this._disabled; }
+  set disabled(value: boolean) {
+    const newValue = coerceBooleanProperty(value);
+    const element = this._elementRef.nativeElement;
+
+    if (this._disabled !== newValue) {
+      this._disabled = newValue;
+      this._disabledChange.emit(newValue);
+    }
+
+    // We need to null check the `blur` method, because it's undefined during SSR.
+    if (newValue && element.blur) {
+      // Normally, native input elements automatically blur if they turn disabled. This behavior
+      // is problematic, because it would mean that it triggers another change detection cycle,
+      // which then causes a changed after checked error if the input element was focused before.
+      element.blur();
+    }
+  }
+  private _disabled: boolean;
+
+  /** Emits when a `change` event is fired on this `<input>`. */
+  @Output() readonly dateChange: EventEmitter<MatDatepickerInputEvent<D>> =
+      new EventEmitter<MatDatepickerInputEvent<D>>();
+
+  /** Emits when an `input` event is fired on this `<input>`. */
+  @Output() readonly dateInput: EventEmitter<MatDatepickerInputEvent<D>> =
+      new EventEmitter<MatDatepickerInputEvent<D>>();
+
+  /** Emits when the value changes (either due to user input or programmatic change). */
+  _valueChange = new EventEmitter<D | null>();
+
+  /** Emits when the disabled state has changed */
+  _disabledChange = new EventEmitter<boolean>();
+
+  _onTouched = () => {};
+
+  private _cvaOnChange: (value: any) => void = () => {};
+  protected _validatorOnChange = () => {};
+  private _valueChangesSubscription = Subscription.EMPTY;
+  private _localeSubscription = Subscription.EMPTY;
+
+  /**
+   * Since the value is kept on the model which is assigned in an Input,
+   * we might get a value before we have a model. This property keeps track
+   * of the value until we have somewhere to assign it.
+   */
+  private _pendingValue: D | null;
+
+  /** The form control validator for whether the input parses. */
+  protected _parseValidator: ValidatorFn = (): ValidationErrors | null => {
+    return this._lastValueValid ?
+        null : {'matDatepickerParse': {'text': this._elementRef.nativeElement.value}};
+  }
+
+  protected _registerModel(model: MatDateSelectionModel<D | null, D>): void {
+    this._model = model;
+    this._valueChangesSubscription.unsubscribe();
+
+    if (this._pendingValue) {
+      this._assignValue(this._pendingValue);
+    }
+
+    this._valueChangesSubscription = this._model.selectionChanged.subscribe(event => {
+      if (event.source !== this) {
+        this._cvaOnChange(event.selection);
+        this._onTouched();
+        this.dateInput.emit(new MatDatepickerInputEvent(this, this._elementRef.nativeElement));
+        this.dateChange.emit(new MatDatepickerInputEvent(this, this._elementRef.nativeElement));
+        this._formatValue(event.selection);
+      }
+    });
+  }
+
+  /** Opens the popup associated with the input. */
+  protected abstract _openPopup(): void;
+
+  /** Assigns a value to the input's model. */
+  protected abstract _assignModelValue(model: D | null): void;
+
+  /** The combined form control validator for this input. */
+  protected abstract _validator: ValidatorFn | null;
+
+  /** Whether the last value set on the input was valid. */
+  protected _lastValueValid = false;
+
+  constructor(
+      protected _elementRef: ElementRef<HTMLInputElement>,
+      @Optional() public _dateAdapter: DateAdapter<D>,
+      @Optional() @Inject(MAT_DATE_FORMATS) private _dateFormats: MatDateFormats) {
+    if (!this._dateAdapter) {
+      throw createMissingDateImplError('DateAdapter');
+    }
+    if (!this._dateFormats) {
+      throw createMissingDateImplError('MAT_DATE_FORMATS');
+    }
+
+    // Update the displayed date when the locale changes.
+    this._localeSubscription = _dateAdapter.localeChanges.subscribe(() => {
+      this.value = this.value;
+    });
+  }
+
+  ngOnDestroy() {
+    this._valueChangesSubscription.unsubscribe();
+    this._localeSubscription.unsubscribe();
+    this._valueChange.complete();
+    this._disabledChange.complete();
+  }
+
+  /** @docs-private */
+  registerOnValidatorChange(fn: () => void): void {
+    this._validatorOnChange = fn;
+  }
+
+  /** @docs-private */
+  validate(c: AbstractControl): ValidationErrors | null {
+    return this._validator ? this._validator(c) : null;
+  }
+
+  // Implemented as part of ControlValueAccessor.
+  writeValue(value: D): void {
+    this.value = value;
+  }
+
+  // Implemented as part of ControlValueAccessor.
+  registerOnChange(fn: (value: any) => void): void {
+    this._cvaOnChange = fn;
+  }
+
+  // Implemented as part of ControlValueAccessor.
+  registerOnTouched(fn: () => void): void {
+    this._onTouched = fn;
+  }
+
+  // Implemented as part of ControlValueAccessor.
+  setDisabledState(isDisabled: boolean): void {
+    this.disabled = isDisabled;
+  }
+
+  _onKeydown(event: KeyboardEvent) {
+    const isAltDownArrow = event.altKey && event.keyCode === DOWN_ARROW;
+
+    if (isAltDownArrow && !this._elementRef.nativeElement.readOnly) {
+      this._openPopup();
+      event.preventDefault();
+    }
+  }
+
+  _onInput(value: string) {
+    const lastValueWasValid = this._lastValueValid;
+    let date = this._dateAdapter.parse(value, this._dateFormats.parse.dateInput);
+    this._lastValueValid = !date || this._dateAdapter.isValid(date);
+    date = this._getValidDateOrNull(date);
+
+    if (!this._dateAdapter.sameDate(date, this.value)) {
+      this._assignValue(date);
+      this._cvaOnChange(date);
+      this._valueChange.emit(date);
+      this.dateInput.emit(new MatDatepickerInputEvent(this, this._elementRef.nativeElement));
+    } else if (lastValueWasValid !== this._lastValueValid) {
+      this._validatorOnChange();
+    }
+  }
+
+  _onChange() {
+    this.dateChange.emit(new MatDatepickerInputEvent(this, this._elementRef.nativeElement));
+  }
+
+  /** Handles blur events on the input. */
+  _onBlur() {
+    // Reformat the input only if we have a valid value.
+    if (this.value) {
+      this._formatValue(this.value);
+    }
+
+    this._onTouched();
+  }
+
+  /** Formats a value and sets it on the input element. */
+  private _formatValue(value: D | null) {
+    this._elementRef.nativeElement.value =
+        value ? this._dateAdapter.format(value, this._dateFormats.display.dateInput) : '';
+  }
+
+  /**
+   * @param obj The object to check.
+   * @returns The given object if it is both a date instance and valid, otherwise null.
+   */
+  protected _getValidDateOrNull(obj: any): D | null {
+    return (this._dateAdapter.isDateInstance(obj) && this._dateAdapter.isValid(obj)) ? obj : null;
+  }
+
+  /** Assigns a value to the model. */
+  private _assignValue(value: D | null) {
+    // We may get some incoming values before the model was
+    // assigned. Save the value so that we can assign it later.
+    if (this._model) {
+      this._assignModelValue(value);
+      this._pendingValue = null;
+    } else {
+      this._pendingValue = value;
+    }
+  }
+
+  // Accept `any` to avoid conflicts with other directives on `<input>` that
+  // may accept different types.
+  static ngAcceptInputType_value: any;
+  static ngAcceptInputType_disabled: BooleanInput;
+}
