+++ src/material/datepicker/public-api.ts
@@ -11,7 +11,12 @@ export * from './calendar';
 export * from './calendar-body';
 export * from './datepicker';
 export * from './datepicker-animations';
-export * from './datepicker-input';
+export {MatDatepickerInputEvent} from './datepicker-input-base';
+export {
+    MAT_DATEPICKER_VALUE_ACCESSOR,
+    MAT_DATEPICKER_VALIDATORS,
+    MatDatepickerInput,
+} from './datepicker-input';
 export * from './datepicker-intl';
 export * from './datepicker-toggle';
 export * from './month-view';
